Refactoring of pupil register, attendance and other bits and pieces.
