+++ admin/controllers/pupil.js
@@ -13,14 +13,20 @@ const schoolDataService = require('../services/data-access/school.data.service')
 const ValidationError = require('../lib/validation-error')
 const winston = require('winston')
 
-const controller = {}
-
-controller.getAddPupil = async (req, res, next, error = null) => {
+/**
+ * Get list of pupils.
+ * @param req
+ * @param res
+ * @param next
+ * @param error
+ * @returns {Promise<void>}
+ */
+const getAddPupil = async (req, res, next, error = null) => {
   res.locals.pageTitle = 'Add pupil'
   try {
-    req.breadcrumbs('Pupil Register', '/school/pupil-register/name/asc')
+    req.breadcrumbs('Pupil Register', '/pupil-register/pupils-list/name/asc')
     req.breadcrumbs(res.locals.pageTitle)
-    res.render('school/add-pupil', {
+    res.render('pupil-register/add-pupil', {
       formData: req.body,
       error: error || new ValidationError(),
       breadcrumbs: req.breadcrumbs()
@@ -30,39 +36,40 @@ controller.getAddPupil = async (req, res, next, error = null) => {
   }
 }
 
-controller.postAddPupil = async (req, res, next) => {
+/**
+ * Post new pupil record.
+ * @param req
+ * @param res
+ * @param next
+ * @returns {Promise<void>}
+ */
+const postAddPupil = async (req, res, next) => {
   res.locals.pageTitle = 'Add pupil'
   try {
-    const school = await schoolDataService.sqlFindOneByDfeNumber(req.user.School)
-    const pupilData = {
-      school_id: school.id,
-      upn: req.body.upn,
-      foreName: req.body.foreName,
-      lastName: req.body.lastName,
-      middleNames: req.body.middleNames,
-      gender: req.body.gender,
-      'dob-month': req.body['dob-month'],
-      'dob-day': req.body['dob-day'],
-      'dob-year': req.body['dob-year']
-    }
-    const pupil = await pupilAddService.addPupil(pupilData)
+    const pupil = await pupilAddService.addPupil(req.body, req.user.schoolId)
     req.flash('info', '1 new pupil has been added')
     const highlight = JSON.stringify([pupil.urlSlug.toString()])
-    res.redirect(`/school/pupil-register/name/asc?hl=${highlight}`)
+    res.redirect(`/pupil-register/pupils-list/name/asc?hl=${highlight}`)
   } catch (error) {
     if (error.name === 'ValidationError') {
-      return controller.getAddPupil(req, res, next, error)
+      return getAddPupil(req, res, next, error)
     }
     winston.warn(error)
     next(error)
   }
 }
 
-controller.getAddMultiplePupils = (req, res, next) => {
+/**
+ * Add multiple pupils, view page.
+ * @param req
+ * @param res
+ * @param next
+ */
+const getAddMultiplePupils = (req, res, next) => {
   res.locals.pageTitle = 'Add multiple pupils'
   const { hasError, fileErrors } = res
   try {
-    req.breadcrumbs('Pupil Register', '/school/pupil-register/name/asc')
+    req.breadcrumbs('Pupil Register', '/pupil-register/pupils-list/name/asc')
     req.breadcrumbs(res.locals.pageTitle)
     res.render('school/add-multiple-pupils', {
       breadcrumbs: req.breadcrumbs(),
@@ -74,7 +81,14 @@ controller.getAddMultiplePupils = (req, res, next) => {
   }
 }
 
-controller.postAddMultiplePupils = async (req, res, next) => {
+/**
+ * Save multiple pupils.
+ * @param req
+ * @param res
+ * @param next
+ * @returns {Promise<*>}
+ */
+const postAddMultiplePupils = async (req, res, next) => {
   let school
   try {
     school = await schoolDataService.sqlFindOneByDfeNumber(req.user.School)
@@ -84,42 +98,52 @@ controller.postAddMultiplePupils = async (req, res, next) => {
   } catch (error) {
     return next(error)
   }
+
   const uploadFile = req.files && req.files.csvTemplateFile
-  const fileErrors = await fileValidator.validate(uploadFile, 'file-upload')
-  if (fileErrors.hasError()) {
-    res.hasError = true
-    res.fileErrors = fileErrors
-    return controller.getAddMultiplePupils(req, res, next)
+  try {
+    const fileErrors = await fileValidator.validate(uploadFile, 'file-upload')
+    if (fileErrors.hasError()) {
+      res.hasError = true
+      res.fileErrors = fileErrors
+      return getAddMultiplePupils(req, res, next)
+    }
+  } catch (error) {
+    return next(error)
   }
+
   let uploadResult
   try {
     uploadResult = await pupilUploadService.upload(school, uploadFile)
   } catch (error) {
     return next(error)
   }
-  // upload error
-  if (uploadResult.error) return next(uploadResult.error)
-  // render with errors
+
+  // Upload errors found
+  if (uploadResult.error) {
+    return next(uploadResult.error)
+  }
+
   if (uploadResult.hasValidationError) {
     req.session.csvErrorFile = uploadResult.csvErrorFile
     res.hasError = uploadResult.hasValidationError
     res.fileErrors = uploadResult.fileErrors
-    return controller.getAddMultiplePupils(req, res, next)
+    return getAddMultiplePupils(req, res, next)
   } else {
     req.flash('info', `${uploadResult.pupilIds && uploadResult.pupilIds.length} new pupils have been added`)
     const savedPupils = await pupilDataService.sqlFindByIds(uploadResult.pupilIds)
     const slugs = savedPupils.map(p => p.urlSlug)
     const qp = encodeURIComponent(JSON.stringify(slugs))
-    res.redirect(`/school/pupil-register/name/asc?hl=${qp}`)
+    res.redirect(`/pupil-register/pupils-list/name/asc?hl=${qp}`)
   }
 }
 
-controller.getAddMultiplePupilsCSVTemplate = async (req, res) => {
-  const file = 'public/csvs/mtc-pupil-details-template-sheet-1.csv'
-  res.download(file)
-}
-
-controller.getErrorCSVFile = async (req, res) => {
+/**
+ * Get error CSV file.
+ * @param req
+ * @param res
+ * @returns {Promise<void>}
+ */
+const getErrorCSVFile = async (req, res) => {
   const blobFile = await azureFileDataService.azureDownloadFile('csvuploads', req.session.csvErrorFile)
   res.setHeader('Content-disposition', 'filename=multiple_pupils_errors.csv')
   res.setHeader('content-type', 'text/csv')
@@ -127,24 +151,31 @@ controller.getErrorCSVFile = async (req, res) => {
   res.end()
 }
 
-controller.getEditPupilById = async (req, res, next) => {
+/**
+ * Get pupil by id.
+ * @param req
+ * @param res
+ * @param next
+ * @returns {Promise<*>}
+ */
+const getEditPupilById = async (req, res, next) => {
   res.locals.pageTitle = 'Edit pupil data'
   try {
     const pupil = await pupilDataService.sqlFindOneBySlug(req.params.id)
     if (!pupil) {
       return next(new Error(`Pupil ${req.params.id} not found`))
     }
+
+    // @TODO: Why is the school data pulled for?
     const school = await schoolDataService.sqlFindOneById(pupil.school_id)
     if (!school) {
       return next(new Error(`School ${pupil.school._id} not found`))
     }
-    const pupilData = R.omit('dateOfBirth', pupil)
-    // expand single date field to 3
-    pupilData['dob-day'] = pupil.dateOfBirth.format('D')
-    pupilData['dob-month'] = pupil.dateOfBirth.format('M')
-    pupilData['dob-year'] = pupil.dateOfBirth.format('YYYY')
+
+    const pupilData = pupilAddService.formatPupilData(pupil)
+
     req.breadcrumbs(res.locals.pageTitle)
-    res.render('school/edit-pupil', {
+    res.render('pupil-register/edit-pupil', {
       formData: pupilData,
       error: new ValidationError(),
       breadcrumbs: req.breadcrumbs()
@@ -154,7 +185,14 @@ controller.getEditPupilById = async (req, res, next) => {
   }
 }
 
-controller.postEditPupil = async (req, res, next) => {
+/**
+ * Post pupil
+ * @param req
+ * @param res
+ * @param next
+ * @returns {Promise<*>}
+ */
+const postEditPupil = async (req, res, next) => {
   let pupil
   let school
   let validationError
@@ -166,10 +204,12 @@ controller.postEditPupil = async (req, res, next) => {
     if (!pupil) {
       return next(new Error(`Pupil ${req.body.urlSlug} not found`))
     }
+
     school = await schoolDataService.sqlFindOneById(pupil.school_id)
     if (!school) {
       return next(new Error(`School not found`))
     }
+
     validationError = await pupilValidator.validate(req.body)
   } catch (error) {
     return next(error)
@@ -177,7 +217,7 @@ controller.postEditPupil = async (req, res, next) => {
 
   if (validationError.hasError()) {
     req.breadcrumbs(res.locals.pageTitle)
-    return res.render('school/edit-pupil', {
+    return res.render('pupil-register/edit-pupil', {
       school,
       formData: req.body,
       error: validationError,
@@ -203,19 +243,22 @@ controller.postEditPupil = async (req, res, next) => {
     next(error)
   }
 
-  // pupil saved - redirect and highlight the saved pupil
   const highlight = JSON.stringify([pupil.urlSlug.toString()])
-  res.redirect(`/school/pupil-register/name/asc?hl=${highlight}`)
+  res.redirect(`/pupil-register/pupils-list/name/asc?hl=${highlight}`)
 }
 
 /**
- * Print the pupil and school pins
+ * Print the pupil and school pins.
+ * @param req
+ * @param res
+ * @param next
+ * @returns {Promise<*>}
  */
-controller.getPrintPupils = async (req, res, next) => {
+const getPrintPupils = async (req, res, next) => {
   res.locals.pageTitle = 'Print pupils'
   try {
     const pupilData = await pupilService.getPrintPupils(req.user.School)
-    res.render('school/pupils-print', {
+    res.render('pupil-register/pupils-print', {
       pupils: pupilData
     })
   } catch (error) {
@@ -223,4 +266,13 @@ controller.getPrintPupils = async (req, res, next) => {
   }
 }
 
-module.exports = controller
+module.exports = {
+  getAddPupil,
+  postAddPupil,
+  getAddMultiplePupils,
+  postAddMultiplePupils,
+  getErrorCSVFile,
+  getEditPupilById,
+  postEditPupil,
+  getPrintPupils
+}
