--- admin/routes/admin.js
@@ -1,11 +1,0 @@
-'use strict'
-
-const express = require('express')
-const router = express.Router()
-const school = require('./school')
-const pupil = require('./pupil')
-
-school(router)
-pupil(router)
-
-module.exports = router
