+++ admin/views/pupil-register/add-pupil.ejs
