+++ admin/views/restart/restart-overview.ejs
@@ -36,7 +36,7 @@
                         <a href="/pupil-pin/generate-pins-overview">Generate pupil PINs</a>
                     </li>
                     <li>
-                        <a href="/school/pupil-register/name/asc" >Pupil Register</a>
+                        <a href="/pupil-register/pupils-list/name/asc" >Pupil Register</a>
                     </li>
                 </ul>
             </nav>
