+++ admin/views/school/school-home.ejs
@@ -11,7 +11,7 @@
 
         <ol class="list">
             <li>
-                <h2 class="heading-small arrow-up"><a href="/school/pupil-register/name/asc">Pupil register</a></h2>
+                <h2 class="heading-small arrow-up"><a href="/pupil-register/pupils-list/name/asc">Pupil register</a></h2>
                 <p>Add, edit and review pupils</p>
             </li>
